tendermint-rpc: Move legacy -> tendermint33
